+++ Changelog.md
@@ -26,7 +26,8 @@ first three params (`TChildProps` can be derived). [#1402](https://github.com/ap
   - Update rollup configurations, refine package exports [#1467](https://github.com/apollographql/react-apollo/pull/1467)
   - Removed unused gzip script [#1468](https://github.com/apollographql/react-apollo/pull/1468)
   - Minify umd and ensure umd name consistency [#1469](https://github.com/apollographql/react-apollo/pull/1469)
-  
+  - Converted `test/test-utils/test-utils.test.js` to `test/test-utils.test.tsx` [#1475](https://github.com/apollographql/react-apollo/pull/1475)
+
 ### 2.0.4
 - rolled back on the lodash-es changes from
   [#1344](https://github.com/apollographql/react-apollo/pull/1344) due to build